Divide cartesianContextProvider
