+++ packages/x-charts/src/context/CartesianProvider/index.ts
@@ -1,0 +1,13 @@
+import { computeValue } from './computeValue';
+import { normalizeAxis } from './normalizeAxis';
+
+export * from './CartesianProvider';
+export * from './CartesianContext';
+export * from './useCartesianContext';
+
+const cartesianProviderUtils = {
+  computeValue,
+  normalizeAxis,
+};
+
+export { cartesianProviderUtils };
