+++ packages/x-charts/src/context/CartesianProvider/useCartesianContext.ts
@@ -1,0 +1,7 @@
+import * as React from 'react';
+import { CartesianContext, CartesianContextState } from './CartesianContext';
+
+export const useCartesianContext = (): CartesianContextState => {
+  const { data } = React.useContext(CartesianContext);
+  return data;
+};
